Only keep v3 as the root page
