+++ redux/selectors.ts
@@ -1,50 +1,70 @@
 import { createSelector } from "reselect";
 import { RootState } from "./store";
+import data from "../data/ingredient_taxonomy.json";
+import { DataType } from "./reducers/editor";
+import { VALUE, QUANTITY, INGREDIENT } from "./reducers/recipes";
 
 export const selectCurrentIngredients = createSelector(
   (state: RootState, currentRecipeId: string) => currentRecipeId,
   (state: RootState) => state.recipe.recipes,
-  (state: RootState) => state.ingredients,
-  (currentRecipeId, recipes, ingredients) =>
-    recipes[currentRecipeId].ingredients.map(({ id, quantity }) => ({
-      ...ingredients.values[id],
-      quantity,
-    }))
+  (currentRecipeId, recipes) => recipes[currentRecipeId].ingredients
 );
 
-export const selectCurrentIngredientsIds = createSelector(
-  (state: RootState, currentRecipeId: string) => currentRecipeId,
-  (state: RootState) => state.recipe.recipes,
-  (state: RootState) => state.ingredients,
-  (currentRecipeId, recipes, ingredients) =>
-    recipes[currentRecipeId].ingredients.map(({ id }) => id)
-);
-
-export const selectAvailableIngredients = createSelector(
-  (state: RootState) => state.ingredients,
-  (state: RootState, ingredientType: string) => ingredientType,
-  (ingredients, ingredientType) =>
-    ingredients.ids
-      .filter((id) => ingredients.values[id].type === ingredientType)
-      .map((id) => ingredients.values[id])
-);
-
-/**
- * Select ingreditent unused in a give recipe, and in a certain category. `ingredientType` can be empty
- */
-export const selectUnusedIngredients = createSelector(
-  (state: RootState) => state.ingredients,
-  (state: RootState, ingredientType: string) => ingredientType,
-  (state: RootState, ingredientType: string, recipeId: string) =>
-    state.recipe.recipes[recipeId].ingredients.map(
-      (ingredient) => ingredient.id
-    ),
-  (ingredients, ingredientType, usedIds) =>
-    ingredients.ids
-      .filter(
-        (id) =>
-          (!ingredientType || ingredients.values[id].type === ingredientType) &&
-          !usedIds.includes(id)
+export const selectURLParams = createSelector(
+  selectCurrentIngredients,
+  (ingredients) => {
+    let i = 0;
+    return ingredients
+      .flatMap((ingredient) =>
+        ingredient.quantities.flatMap((quantity) => {
+          i += 1;
+          return [
+            `${INGREDIENT}${i}=${ingredient.id}`,
+            `${QUANTITY}${i}=${quantity.id}`,
+            `${VALUE}${i}=${quantity.value}`,
+          ];
+        })
       )
-      .map((id) => ingredients.values[id])
+      .join("&");
+  }
+);
+
+export const selectEditorState = createSelector(
+  (state: RootState) => state.editor,
+  (editor) => editor
+);
+
+export const selectEditorView = createSelector(
+  (state: RootState) => state.editor,
+  (editor) => editor.currentView
+);
+
+export const selectEditorCurrentType = createSelector(
+  (state: RootState) => state.editor,
+  (editor) =>
+    (data as DataType).find(
+      (type) => type["Ingredient type id"] === editor.typeId
+    ) ?? null
+);
+
+export const selectEditorCurrentIngredient = createSelector(
+  (state: RootState) => state.editor,
+  selectEditorCurrentType,
+  (editor, currentType) =>
+    (currentType &&
+      currentType.ingredients.find(
+        (ingredient) => ingredient["Ingredient id"] === editor.ingredientId
+      )) ??
+    null
+);
+
+export const selectEditorCurrentQuantity = createSelector(
+  (state: RootState) => state.editor,
+  selectEditorCurrentIngredient,
+  (editor, currentIngredient) =>
+    (currentIngredient &&
+      currentIngredient.quantities.find(
+        (quantity) => quantity["Quantity id"] === editor.quantityId
+      )) ??
+    null
 );
